linkedin-agent-django/requirements.txt
# Django
Django>=4.2
djangorestframework>=3.14

# CORS support for Chrome Extension
django-cors-headers>=4.0

# Database
psycopg2-binary>=2.9  # PostgreSQL adapter

# Environment variables
python-dotenv>=1.0

# AI
openai>=1.0

# Optional: for DATABASE_URL parsing
dj-database-url>=2.0
